+++ packages/tui/CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
 ## [Unreleased]
+
 ### Added
 
 - Added `renderInlineMarkdown()` function to render inline markdown (bold, italic, code, links, strikethrough) to styled strings
+
+### Fixed
+
+- Fixed editor consuming user-rebound copy keys, preventing custom keybindings from working in the editor
 
 ## [13.14.1] - 2026-03-21
 ### Added